Refactoring SalesforceSource class for better code organization and testability
- In the current form, it's very difficult to add tests for changes as we need to handle all sorts of calls.
- Moved the code related to histogram calculation to a new class - SalesforceHistogramService. This allows us to mock all that happens in this class, making writing tests easy.
- Added a test for the new change introduced as part of https://github.com/apache/gobblin/pull/3766
