+++ gobblin-salesforce/src/main/java/org/apache/gobblin/salesforce/SalesforceSource.java
@@ -91,35 +91,23 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
   public static final String USE_ALL_OBJECTS = "use.all.objects";
   public static final boolean DEFAULT_USE_ALL_OBJECTS = false;
 
-  private static final String ENABLE_DYNAMIC_PROBING = "salesforce.enableDynamicProbing";
-  private static final String DYNAMIC_PROBING_LIMIT = "salesforce.dynamicProbingLimit";
-  private static final int DEFAULT_DYNAMIC_PROBING_LIMIT = 1000;
-  private static final String MIN_TARGET_PARTITION_SIZE = "salesforce.minTargetPartitionSize";
-  private static final int DEFAULT_MIN_TARGET_PARTITION_SIZE = 250000;
-  // this is used to generate histogram buckets smaller than the target partition size to allow for more even
-  // packing of the generated partitions
-  private static final String PROBE_TARGET_RATIO = "salesforce.probeTargetRatio";
-  private static final double DEFAULT_PROBE_TARGET_RATIO = 0.60;
-  private static final int MIN_SPLIT_TIME_MILLIS = 1000;
+  @VisibleForTesting
+  static final String ENABLE_DYNAMIC_PROBING = "salesforce.enableDynamicProbing";
+  static final String MIN_TARGET_PARTITION_SIZE = "salesforce.minTargetPartitionSize";
+  static final int DEFAULT_MIN_TARGET_PARTITION_SIZE = 250000;
 
-  private static final String DAY_PARTITION_QUERY_TEMPLATE =
-      "SELECT count(${column}) cnt, DAY_ONLY(${column}) time FROM ${table} " + "WHERE ${column} ${greater} ${start}"
-          + " AND ${column} ${less} ${end} GROUP BY DAY_ONLY(${column}) ORDER BY DAY_ONLY(${column})";
-  private static final String PROBE_PARTITION_QUERY_TEMPLATE = "SELECT count(${column}) cnt FROM ${table} "
-      + "WHERE ${column} ${greater} ${start} AND ${column} ${less} ${end}";
-
-  private static final String ENABLE_DYNAMIC_PARTITIONING = "salesforce.enableDynamicPartitioning";
-  private static final String EARLY_STOP_TOTAL_RECORDS_LIMIT = "salesforce.earlyStopTotalRecordsLimit";
+  @VisibleForTesting
+  static final String ENABLE_DYNAMIC_PARTITIONING = "salesforce.enableDynamicPartitioning";
+  @VisibleForTesting
+  static final String EARLY_STOP_TOTAL_RECORDS_LIMIT = "salesforce.earlyStopTotalRecordsLimit";
   private static final long DEFAULT_EARLY_STOP_TOTAL_RECORDS_LIMIT = DEFAULT_MIN_TARGET_PARTITION_SIZE * 4;
 
-  private static final String SECONDS_FORMAT = "yyyy-MM-dd-HH:mm:ss";
-  private static final String ZERO_TIME_SUFFIX = "-00:00:00";
+  static final String SECONDS_FORMAT = "yyyy-MM-dd-HH:mm:ss";
 
-  private static final Gson GSON = new Gson();
   private boolean isEarlyStopped = false;
   protected SalesforceConnector salesforceConnector = null;
 
-  private SfConfig workUnitConf;
+  private SalesforceHistogramService salesforceHistogramService;
 
   public SalesforceSource() {
     this.lineageInfo = Optional.absent();
@@ -128,6 +116,12 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
   @VisibleForTesting
   SalesforceSource(LineageInfo lineageInfo) {
     this.lineageInfo = Optional.fromNullable(lineageInfo);
+  }
+
+  @VisibleForTesting
+  SalesforceSource(SalesforceHistogramService salesforceHistogramService) {
+    this.lineageInfo = Optional.absent();
+    this.salesforceHistogramService = salesforceHistogramService;
   }
 
   @Override
@@ -155,8 +149,12 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
   }
   @Override
   protected List<WorkUnit> generateWorkUnits(SourceEntity sourceEntity, SourceState state, long previousWatermark) {
-    List<WorkUnit> workUnits = null;
-    workUnitConf = new SfConfig(state.getProperties());
+    SfConfig sfConfig = new SfConfig(state.getProperties());
+    if (salesforceHistogramService == null) {
+      salesforceHistogramService = new SalesforceHistogramService(sfConfig);
+    }
+
+    List<WorkUnit> workUnits;
     String partitionType = state.getProp(SALESFORCE_PARTITION_TYPE, "");
     if (partitionType.equals("PK_CHUNKING")) {
       // pk-chunking only supports start-time by source.querybased.start.value, and does not support end-time.
@@ -164,10 +162,10 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
       // we should only pk chunking based work units only in case of snapshot/full ingestion
       workUnits = generateWorkUnitsPkChunking(sourceEntity, state, previousWatermark);
     } else {
-      workUnits = generateWorkUnitsStrategy(sourceEntity, state, previousWatermark);
+      workUnits = generateWorkUnitsHelper(sourceEntity, state, previousWatermark);
     }
     log.info("====Generated {} workUnit(s)====", workUnits.size());
-    if (workUnitConf.partitionOnly) {
+    if (sfConfig.partitionOnly) {
       log.info("It is partitionOnly mode, return blank workUnit list");
       return new ArrayList<>();
     } else {
@@ -277,11 +275,13 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
   }
 
   /**
-   *
+   * Generates {@link WorkUnit}s based on a bunch of config values like max number of partitions, early stop,
+   * dynamic partitioning, dynamic probing, etc.
    */
-  private List<WorkUnit> generateWorkUnitsStrategy(SourceEntity sourceEntity, SourceState state, long previousWatermark) {
-    Boolean disableSoft = state.getPropAsBoolean(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, false);
-    log.info("disable soft delete pull: " + disableSoft);
+  @VisibleForTesting
+  List<WorkUnit> generateWorkUnitsHelper(SourceEntity sourceEntity, SourceState state, long previousWatermark) {
+    boolean isSoftDeletePullDisabled = state.getPropAsBoolean(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, false);
+    log.info("disable soft delete pull: " + isSoftDeletePullDisabled);
     WatermarkType watermarkType = WatermarkType.valueOf(
         state.getProp(ConfigurationKeys.SOURCE_QUERYBASED_WATERMARK_TYPE, ConfigurationKeys.DEFAULT_WATERMARK_TYPE)
             .toUpperCase());
@@ -292,10 +292,12 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
     int minTargetPartitionSize = state.getPropAsInt(MIN_TARGET_PARTITION_SIZE, DEFAULT_MIN_TARGET_PARTITION_SIZE);
 
     // Only support time related watermark
-    if (watermarkType == WatermarkType.SIMPLE || Strings.isNullOrEmpty(watermarkColumn) || !state.getPropAsBoolean(
-        ENABLE_DYNAMIC_PARTITIONING)) {
+    if (watermarkType == WatermarkType.SIMPLE
+        || Strings.isNullOrEmpty(watermarkColumn)
+        || !state.getPropAsBoolean(ENABLE_DYNAMIC_PARTITIONING)) {
       List<WorkUnit> workUnits = super.generateWorkUnits(sourceEntity, state, previousWatermark);
-      workUnits.forEach(x -> x.setProp(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, disableSoft));
+      workUnits.forEach(workUnit ->
+          workUnit.setProp(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, isSoftDeletePullDisabled));
       return workUnits;
     }
 
@@ -305,7 +307,8 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
     }
 
     Partition partition = partitioner.getGlobalPartition(previousWatermark);
-    Histogram histogram = getHistogram(sourceEntity.getSourceEntityName(), watermarkColumn, state, partition);
+    Histogram histogram =
+        salesforceHistogramService.getHistogram(sourceEntity.getSourceEntityName(), watermarkColumn, state, partition);
 
     // we should look if the count is too big, cut off early if count exceeds the limit, or bucket size is too large
 
@@ -316,8 +319,8 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
       histogramAdjust = new Histogram();
       for (HistogramGroup group : histogram.getGroups()) {
         histogramAdjust.add(group);
-        if (histogramAdjust.getTotalRecordCount() > state
-            .getPropAsLong(EARLY_STOP_TOTAL_RECORDS_LIMIT, DEFAULT_EARLY_STOP_TOTAL_RECORDS_LIMIT)) {
+        long earlyStopRecordLimit = state.getPropAsLong(EARLY_STOP_TOTAL_RECORDS_LIMIT, DEFAULT_EARLY_STOP_TOTAL_RECORDS_LIMIT);
+        if (histogramAdjust.getTotalRecordCount() > earlyStopRecordLimit) {
           break;
         }
       }
@@ -345,7 +348,8 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
     state.setProp(Partitioner.IS_EARLY_STOPPED, isEarlyStopped);
 
     List<WorkUnit> workUnits = super.generateWorkUnits(sourceEntity, state, previousWatermark);
-    workUnits.stream().forEach(x -> x.setProp(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, disableSoft));
+    workUnits.forEach(workUnit ->
+        workUnit.setProp(SOURCE_QUERYBASED_SALESFORCE_IS_SOFT_DELETES_PULL_DISABLED, isSoftDeletePullDisabled));
     return workUnits;
   }
 
@@ -353,12 +357,13 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
     return state.getPropAsBoolean(ConfigurationKeys.SOURCE_EARLY_STOP_ENABLED, ConfigurationKeys.DEFAULT_SOURCE_EARLY_STOP_ENABLED);
   }
 
+  @VisibleForTesting
   String generateSpecifiedPartitions(Histogram histogram, int minTargetPartitionSize, int maxPartitions, long lowWatermark,
       long expectedHighWatermark) {
     int interval = computeTargetPartitionSize(histogram, minTargetPartitionSize, maxPartitions);
     int totalGroups = histogram.getGroups().size();
 
-    log.info("Histogram total record count: " + histogram.totalRecordCount);
+    log.info("Histogram total record count: " + histogram.getTotalRecordCount());
     log.info("Histogram total groups: " + totalGroups);
     log.info("maxPartitions: " + maxPartitions);
     log.info("interval: " + interval);
@@ -384,15 +389,15 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
        * will have total size larger or equal to interval x 2. Hence, we are saturating all intervals (with original size)
        * without leaving any unused space in between. We could choose x3,x4... but it is not space efficient.
        */
-      if (count != 0 && count + group.count >= 2 * interval) {
+      if (count != 0 && count + group.getCount() >= 2 * interval) {
         // Summarize current group
         statistics.addValue(count);
         // A step-in start
         partitionPoints.add(Utils.toDateTimeFormat(group.getKey(), SECONDS_FORMAT, Partitioner.WATERMARKTIMEFORMAT));
-        count = group.count;
+        count = group.getCount();
       } else {
         // Add group into current partition
-        count += group.count;
+        count += group.getCount();
       }
 
       if (count >= interval) {
@@ -428,325 +433,7 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
    */
   private int computeTargetPartitionSize(Histogram histogram, int minTargetPartitionSize, int maxPartitions) {
     return Math.max(minTargetPartitionSize,
-        DoubleMath.roundToInt((double) histogram.totalRecordCount / maxPartitions, RoundingMode.CEILING));
-  }
-
-  /**
-   * Get a {@link JsonArray} containing the query results
-   */
-  @SneakyThrows
-  private JsonArray getRecordsForQuery(SalesforceConnector connector, String query) {
-    RestApiProcessingException exception = null;
-    for (int i = 0; i < workUnitConf.restApiRetryLimit + 1; i++) {
-      try {
-        String soqlQuery = SalesforceExtractor.getSoqlUrl(query);
-        List<Command> commands = RestApiConnector.constructGetCommand(connector.getFullUri(soqlQuery));
-        CommandOutput<?, ?> response = connector.getResponse(commands);
-
-        String output;
-        Iterator<String> itr = (Iterator<String>) response.getResults().values().iterator();
-        if (itr.hasNext()) {
-          output = itr.next();
-        } else {
-          throw new DataRecordException("Failed to get data from salesforce; REST response has no output");
-        }
-
-        return GSON.fromJson(output, JsonObject.class).getAsJsonArray("records");
-      } catch (RestApiClientException | DataRecordException e) {
-        throw new RuntimeException("Fail to get data from salesforce", e);
-      } catch (RestApiProcessingException e) {
-        exception = e;
-        log.info("Caught RestApiProcessingException, retrying({}) rest query: {}", i+1, query);
-        Thread.sleep(workUnitConf.restApiRetryInterval);
-        continue;
-      }
-    }
-    throw new RuntimeException("Fail to get data from salesforce", exception);
-  }
-
-  /**
-   * Get the row count for a time range
-   */
-  private int getCountForRange(TableCountProbingContext probingContext, StrSubstitutor sub,
-      Map<String, String> subValues, long startTime, long endTime) {
-    String startTimeStr = Utils.dateToString(new Date(startTime), SalesforceExtractor.SALESFORCE_TIMESTAMP_FORMAT);
-    String endTimeStr = Utils.dateToString(new Date(endTime), SalesforceExtractor.SALESFORCE_TIMESTAMP_FORMAT);
-
-    subValues.put("start", startTimeStr);
-    subValues.put("end", endTimeStr);
-
-    String query = sub.replace(PROBE_PARTITION_QUERY_TEMPLATE);
-
-    log.debug("Count query: " + query);
-    probingContext.probeCount++;
-
-    JsonArray records = getRecordsForQuery(probingContext.connector, query);
-    Iterator<JsonElement> elements = records.iterator();
-    JsonObject element = elements.next().getAsJsonObject();
-
-    return element.get("cnt").getAsInt();
-  }
-
-  /**
-   * Split a histogram bucket along the midpoint if it is larger than the bucket size limit.
-   */
-  private void getHistogramRecursively(TableCountProbingContext probingContext, Histogram histogram, StrSubstitutor sub,
-      Map<String, String> values, int count, long startEpoch, long endEpoch) {
-    long midpointEpoch = startEpoch + (endEpoch - startEpoch) / 2;
-
-    // don't split further if small, above the probe limit, or less than 1 second difference between the midpoint and start
-    if (count <= probingContext.bucketSizeLimit
-        || probingContext.probeCount > probingContext.probeLimit
-        || (midpointEpoch - startEpoch < MIN_SPLIT_TIME_MILLIS)) {
-      histogram.add(new HistogramGroup(Utils.epochToDate(startEpoch, SECONDS_FORMAT), count));
-      return;
-    }
-
-    int countLeft = getCountForRange(probingContext, sub, values, startEpoch, midpointEpoch);
-
-    getHistogramRecursively(probingContext, histogram, sub, values, countLeft, startEpoch, midpointEpoch);
-    log.debug("Count {} for left partition {} to {}", countLeft, startEpoch, midpointEpoch);
-
-    int countRight = count - countLeft;
-
-    getHistogramRecursively(probingContext, histogram, sub, values, countRight, midpointEpoch, endEpoch);
-    log.debug("Count {} for right partition {} to {}", countRight, midpointEpoch, endEpoch);
-  }
-
-  /**
-   * Get a histogram for the time range by probing to break down large buckets. Use count instead of
-   * querying if it is non-negative.
-   */
-  private Histogram getHistogramByProbing(TableCountProbingContext probingContext, int count, long startEpoch,
-      long endEpoch) {
-    Histogram histogram = new Histogram();
-
-    Map<String, String> values = new HashMap<>();
-    values.put("table", probingContext.entity);
-    values.put("column", probingContext.watermarkColumn);
-    values.put("greater", ">=");
-    values.put("less", "<");
-    StrSubstitutor sub = new StrSubstitutor(values);
-
-    getHistogramRecursively(probingContext, histogram, sub, values, count, startEpoch, endEpoch);
-
-    return histogram;
-  }
-
-  /**
-   * Refine the histogram by probing to split large buckets
-   * @return the refined histogram
-   */
-  private Histogram getRefinedHistogram(SalesforceConnector connector, String entity, String watermarkColumn,
-      SourceState state, Partition partition, Histogram histogram) {
-    final int maxPartitions = state.getPropAsInt(SOURCE_MAX_NUMBER_OF_PARTITIONS, DEFAULT_MAX_NUMBER_OF_PARTITIONS);
-    final int probeLimit = state.getPropAsInt(DYNAMIC_PROBING_LIMIT, DEFAULT_DYNAMIC_PROBING_LIMIT);
-    final int minTargetPartitionSize = state.getPropAsInt(MIN_TARGET_PARTITION_SIZE, DEFAULT_MIN_TARGET_PARTITION_SIZE);
-    final Histogram outputHistogram = new Histogram();
-    final double probeTargetRatio = state.getPropAsDouble(PROBE_TARGET_RATIO, DEFAULT_PROBE_TARGET_RATIO);
-    final int bucketSizeLimit =
-        (int) (probeTargetRatio * computeTargetPartitionSize(histogram, minTargetPartitionSize, maxPartitions));
-
-    log.info("Refining histogram with bucket size limit {}.", bucketSizeLimit);
-
-    HistogramGroup currentGroup;
-    HistogramGroup nextGroup;
-    final TableCountProbingContext probingContext =
-        new TableCountProbingContext(connector, entity, watermarkColumn, bucketSizeLimit, probeLimit);
-
-    if (histogram.getGroups().isEmpty()) {
-      return outputHistogram;
-    }
-
-    // make a copy of the histogram list and add a dummy entry at the end to avoid special processing of the last group
-    List<HistogramGroup> list = new ArrayList(histogram.getGroups());
-    Date hwmDate = Utils.toDate(partition.getHighWatermark(), Partitioner.WATERMARKTIMEFORMAT);
-    list.add(new HistogramGroup(Utils.epochToDate(hwmDate.getTime(), SECONDS_FORMAT), 0));
-
-    for (int i = 0; i < list.size() - 1; i++) {
-      currentGroup = list.get(i);
-      nextGroup = list.get(i + 1);
-
-      // split the group if it is larger than the bucket size limit
-      if (currentGroup.count > bucketSizeLimit) {
-        long startEpoch = Utils.toDate(currentGroup.getKey(), SECONDS_FORMAT).getTime();
-        long endEpoch = Utils.toDate(nextGroup.getKey(), SECONDS_FORMAT).getTime();
-
-        outputHistogram.add(getHistogramByProbing(probingContext, currentGroup.count, startEpoch, endEpoch));
-      } else {
-        outputHistogram.add(currentGroup);
-      }
-    }
-
-    log.info("Executed {} probes for refining the histogram.", probingContext.probeCount);
-
-    // if the probe limit has been reached then print a warning
-    if (probingContext.probeCount >= probingContext.probeLimit) {
-      log.warn("Reached the probe limit");
-    }
-
-    return outputHistogram;
-  }
-
-  /**
-   * Get a histogram with day granularity buckets.
-   */
-  private Histogram getHistogramByDayBucketing(SalesforceConnector connector, String entity, String watermarkColumn,
-      Partition partition) {
-    Histogram histogram = new Histogram();
-
-    Calendar calendar = new GregorianCalendar();
-    Date startDate = Utils.toDate(partition.getLowWatermark(), Partitioner.WATERMARKTIMEFORMAT);
-    calendar.setTime(startDate);
-    int startYear = calendar.get(Calendar.YEAR);
-    String lowWatermarkDate = Utils.dateToString(startDate, SalesforceExtractor.SALESFORCE_TIMESTAMP_FORMAT);
-
-    Date endDate = Utils.toDate(partition.getHighWatermark(), Partitioner.WATERMARKTIMEFORMAT);
-    calendar.setTime(endDate);
-    int endYear = calendar.get(Calendar.YEAR);
-    String highWatermarkDate = Utils.dateToString(endDate, SalesforceExtractor.SALESFORCE_TIMESTAMP_FORMAT);
-
-    Map<String, String> values = new HashMap<>();
-    values.put("table", entity);
-    values.put("column", watermarkColumn);
-    StrSubstitutor sub = new StrSubstitutor(values);
-
-    for (int year = startYear; year <= endYear; year++) {
-      if (year == startYear) {
-        values.put("start", lowWatermarkDate);
-        values.put("greater", partition.isLowWatermarkInclusive() ? ">=" : ">");
-      } else {
-        values.put("start", getDateString(year));
-        values.put("greater", ">=");
-      }
-
-      if (year == endYear) {
-        values.put("end", highWatermarkDate);
-        values.put("less", partition.isHighWatermarkInclusive() ? "<=" : "<");
-      } else {
-        values.put("end", getDateString(year + 1));
-        values.put("less", "<");
-      }
-
-      String query = sub.replace(DAY_PARTITION_QUERY_TEMPLATE);
-      log.info("Histogram query: " + query);
-
-      histogram.add(parseDayBucketingHistogram(getRecordsForQuery(connector, query)));
-    }
-
-    return histogram;
-  }
-
-  protected SalesforceConnector getConnector(State state) {
-    if (this.salesforceConnector == null) {
-      this.salesforceConnector = new SalesforceConnector(state);
-    }
-    return this.salesforceConnector;
-  }
-
-  /**
-   * Generate the histogram
-   */
-  private Histogram getHistogram(String entity, String watermarkColumn, SourceState state,
-      Partition partition) {
-    SalesforceConnector connector = getConnector(state);
-
-    try {
-      if (!connector.connect()) {
-        throw new RuntimeException("Failed to connect.");
-      }
-    } catch (RestApiConnectionException e) {
-      throw new RuntimeException("Failed to connect.", e);
-    }
-
-    Histogram histogram = getHistogramByDayBucketing(connector, entity, watermarkColumn, partition);
-
-    // exchange the first histogram group key with the global low watermark to ensure that the low watermark is captured
-    // in the range of generated partitions
-    HistogramGroup firstGroup = histogram.get(0);
-    Date lwmDate = Utils.toDate(partition.getLowWatermark(), Partitioner.WATERMARKTIMEFORMAT);
-    histogram.getGroups().set(0, new HistogramGroup(Utils.epochToDate(lwmDate.getTime(), SECONDS_FORMAT),
-        firstGroup.getCount()));
-
-    // refine the histogram
-    if (state.getPropAsBoolean(ENABLE_DYNAMIC_PROBING)) {
-      histogram = getRefinedHistogram(connector, entity, watermarkColumn, state, partition, histogram);
-    }
-
-    return histogram;
-  }
-
-  private String getDateString(int year) {
-    Calendar calendar = new GregorianCalendar();
-    calendar.clear();
-    calendar.set(Calendar.YEAR, year);
-    return Utils.dateToString(calendar.getTime(), SalesforceExtractor.SALESFORCE_TIMESTAMP_FORMAT);
-  }
-
-  /**
-   * Parse the query results into a {@link Histogram}
-   */
-  private Histogram parseDayBucketingHistogram(JsonArray records) {
-    log.info("Parse day-based histogram");
-
-    Histogram histogram = new Histogram();
-
-    Iterator<JsonElement> elements = records.iterator();
-    JsonObject element;
-
-    while (elements.hasNext()) {
-      element = elements.next().getAsJsonObject();
-      String time = element.get("time").getAsString() + ZERO_TIME_SUFFIX;
-      int count = element.get("cnt").getAsInt();
-
-      histogram.add(new HistogramGroup(time, count));
-    }
-
-    return histogram;
-  }
-
-  @AllArgsConstructor
-  static class HistogramGroup {
-    @Getter
-    private final String key;
-    @Getter
-    private final int count;
-
-    @Override
-    public String toString() {
-      return key + ":" + count;
-    }
-  }
-
-  static class Histogram {
-    @Getter
-    private long totalRecordCount;
-    @Getter
-    private List<HistogramGroup> groups;
-
-    Histogram() {
-      totalRecordCount = 0;
-      groups = new ArrayList<>();
-    }
-
-    void add(HistogramGroup group) {
-      groups.add(group);
-      totalRecordCount += group.count;
-    }
-
-    void add(Histogram histogram) {
-      groups.addAll(histogram.getGroups());
-      totalRecordCount += histogram.totalRecordCount;
-    }
-
-    HistogramGroup get(int idx) {
-      return this.groups.get(idx);
-    }
-
-    @Override
-    public String toString() {
-      return groups.toString();
-    }
+        DoubleMath.roundToInt((double) histogram.getTotalRecordCount() / maxPartitions, RoundingMode.CEILING));
   }
 
   protected Set<SourceEntity> getSourceEntities(State state) {
@@ -788,17 +475,10 @@ public class SalesforceSource extends QueryBasedSource<JsonArray, JsonElement> {
     return result;
   }
 
-  /**
-   * Context for probing the table for row counts of a time range
-   */
-  @RequiredArgsConstructor
-  private static class TableCountProbingContext {
-    private final SalesforceConnector connector;
-    private final String entity;
-    private final String watermarkColumn;
-    private final int bucketSizeLimit;
-    private final int probeLimit;
-
-    private int probeCount = 0;
+  SalesforceConnector getConnector(State state) {
+    if (this.salesforceConnector == null) {
+      this.salesforceConnector = new SalesforceConnector(state);
+    }
+    return this.salesforceConnector;
   }
 }
